Generalize the progress callback to the core Atlas fetches
The download in BrainGlobeAtlas was the only path reporting bytes to
fn_update, but core.Atlas lazily fetches the template, annotation,
hemispheres and additional references from the same store, each with its own
TqdmCallback, so a caller passing a handler saw nothing for those.

Move _ProgressCallback and _download_callback into core, give core.Atlas an
optional fn_update, and use the shared callback at all five lazy fetches.
BrainGlobeAtlas now hands its handler to super().__init__ rather than keeping
it to itself.
