+++ brainglobe_atlasapi/bg_atlas.py
@@ -7,12 +7,12 @@ from pathlib import Path
 from typing import List, Optional, Tuple, Union
 
 import s3fs
-from fsspec.callbacks import Callback, TqdmCallback
 from rich import print as rprint
 from rich.console import Console
 
 from brainglobe_atlasapi import config, core
 from brainglobe_atlasapi.atlas_name import AtlasName
+from brainglobe_atlasapi.core import _download_callback
 from brainglobe_atlasapi.descriptors import (
     V3_ANNOTATION_MAP_NAME,
     V3_ANNOTATION_MASKS_NAME,
@@ -38,58 +38,6 @@ def _version_tuple_from_str(version_str):
 
 def _version_str_from_tuple(version_tuple: Tuple[int, ...]) -> str:
     return "_".join(str(num) for num in version_tuple)
-
-
-class _ProgressCallback(TqdmCallback):
-    """Show a tqdm bar and report progress to an external handler.
-
-    ``fsspec`` drives the callback passed to ``get`` with a file count, and
-    asks it for a per-file callback through ``branched`` which is driven with
-    a byte count. ``fn_update`` is documented to take completed and total
-    bytes, so it is attached to the branched callbacks.
-
-    Parameters
-    ----------
-    fn_update : Callable
-        Handler called as ``fn_update(completed, total)`` with the bytes
-        transferred so far and the total size of the file being fetched.
-    """
-
-    def __init__(self, fn_update: Callable[[int, int], None], **kwargs):
-        super().__init__(**kwargs)
-        self.fn_update = fn_update
-
-    def branched(self, path_1, path_2, **kwargs):
-        """Return a per-file callback that reports bytes to ``fn_update``."""
-        kwargs["callback"] = Callback(
-            hooks={
-                "fn_update": lambda size, value, **_: self.fn_update(
-                    value, size or 0
-                )
-            }
-        )
-        return super().branched(path_1, path_2, **kwargs)
-
-
-def _download_callback(
-    fn_update: Optional[Callable[[int, int], None]],
-) -> TqdmCallback:
-    """Build the fsspec callback used for a single atlas download step.
-
-    Parameters
-    ----------
-    fn_update : Callable, optional
-        Handler to report download progress to. If None, only the tqdm
-        progress bar is shown.
-
-    Returns
-    -------
-    TqdmCallback
-        The callback to pass to ``fsspec``.
-    """
-    if fn_update is None:
-        return TqdmCallback()
-    return _ProgressCallback(fn_update)
 
 
 class BrainGlobeAtlas(core.Atlas):
@@ -165,7 +113,9 @@ class BrainGlobeAtlas(core.Atlas):
                     "download."
                 )
 
-        super().__init__(self.brainglobe_dir / self.local_full_name)
+        super().__init__(
+            self.brainglobe_dir / self.local_full_name, fn_update=fn_update
+        )
 
         if check_latest:
             self.check_latest_version()
