+++ brainglobe_atlasapi/core.py
@@ -2,10 +2,12 @@
 
 import warnings
 from collections import UserDict, deque
+from collections.abc import Callable
 from pathlib import Path
 from typing import (
     Dict,
     List,
+    Optional,
     Tuple,
     Union,
 )
@@ -17,7 +19,7 @@ import pandas as pd
 import s3fs
 import zarr
 from brainglobe_space import AnatomicalSpace
-from fsspec.callbacks import TqdmCallback
+from fsspec.callbacks import Callback, TqdmCallback
 
 from brainglobe_atlasapi.descriptors import (
     ANNOTATION_DTYPE,
@@ -54,6 +56,58 @@ def _determine_pyramid_level(
     raise ValueError(f"Requested resolution {resolution} um is invalid.")
 
 
+class _ProgressCallback(TqdmCallback):
+    """Show a tqdm bar and report progress to an external handler.
+
+    ``fsspec`` drives the callback passed to ``get`` with a file count, and
+    asks it for a per-file callback through ``branched`` which is driven with
+    a byte count. ``fn_update`` is documented to take completed and total
+    bytes, so it is attached to the branched callbacks.
+
+    Parameters
+    ----------
+    fn_update : Callable
+        Handler called as ``fn_update(completed, total)`` with the bytes
+        transferred so far and the total size of the file being fetched.
+    """
+
+    def __init__(self, fn_update: Callable[[int, int], None], **kwargs):
+        super().__init__(**kwargs)
+        self.fn_update = fn_update
+
+    def branched(self, path_1, path_2, **kwargs):
+        """Return a per-file callback that reports bytes to ``fn_update``."""
+        kwargs["callback"] = Callback(
+            hooks={
+                "fn_update": lambda size, value, **_: self.fn_update(
+                    value, size or 0
+                )
+            }
+        )
+        return super().branched(path_1, path_2, **kwargs)
+
+
+def _download_callback(
+    fn_update: Optional[Callable[[int, int], None]],
+) -> TqdmCallback:
+    """Build the fsspec callback used for a single atlas download step.
+
+    Parameters
+    ----------
+    fn_update : Callable, optional
+        Handler to report download progress to. If None, only the tqdm
+        progress bar is shown.
+
+    Returns
+    -------
+    TqdmCallback
+        The callback to pass to ``fsspec``.
+    """
+    if fn_update is None:
+        return TqdmCallback()
+    return _ProgressCallback(fn_update)
+
+
 class Atlas:
     """Base class to handle atlases in BrainGlobe.
 
@@ -66,7 +120,11 @@ class Atlas:
     left_hemisphere_value = 1
     right_hemisphere_value = 2
 
-    def __init__(self, path):
+    def __init__(self, path, fn_update: Optional[Callable] = None):
+        # Progress handler shared by every lazy fetch below, so a caller that
+        # passes one sees byte progress for the pyramid levels too and not only
+        # for the initial download.
+        self.fn_update = fn_update
         self._template_pyramid_level = 0
         self._annotation_pyramid_level = 0
         self.fs = s3fs.S3FileSystem(anon=True)
@@ -125,7 +183,7 @@ class Atlas:
 
         # Add entry for file paths:
         for struct in structures_list:
-            struct["mesh_filename"] = meshes_path / f'{struct["id"]}'
+            struct["mesh_filename"] = meshes_path / f"{struct['id']}"
 
         self.structures = StructuresDict(structures_list)
 
@@ -227,7 +285,7 @@ class Atlas:
                 remote_path,
                 resolution_path,
                 recursive=True,
-                callback=TqdmCallback(),
+                callback=_download_callback(self.fn_update),
             )
 
         self._template = multiscale.images[
@@ -275,7 +333,7 @@ class Atlas:
                 remote_path,
                 resolution_path,
                 recursive=True,
-                callback=TqdmCallback(),
+                callback=_download_callback(self.fn_update),
             )
 
         self._annotation = multiscale.images[
@@ -336,7 +394,7 @@ class Atlas:
                     remote_path,
                     resolution_path,
                     recursive=True,
-                    callback=TqdmCallback(),
+                    callback=_download_callback(self.fn_update),
                 )
 
             self._hemispheres = multiscale.images[
@@ -657,7 +715,7 @@ class Atlas:
                 )
             except IndexError:
                 raise ValueError(
-                    f'Structure {self.structures[structure]["acronym"]} '
+                    f"Structure {self.structures[structure]['acronym']} "
                     f"has no descendants at hierarchy level {hierarchy_level}"
                 )
 
@@ -722,7 +780,7 @@ class Atlas:
                     remote_path,
                     local_path,
                     recursive=True,
-                    callback=TqdmCallback(),
+                    callback=_download_callback(self.fn_update),
                 )
             except FileNotFoundError as e:
                 raise FileNotFoundError(
@@ -817,7 +875,7 @@ class AdditionalRefDict(UserDict):
                     remote_path,
                     resolution_path,
                     recursive=True,
-                    callback=TqdmCallback(),
+                    callback=_download_callback(self.fn_update),
                 )
             self.data[key] = multiscale.images[pyramid_level].data.compute()
